Separate ADT analysis implementation from lowering implementation
